refactor: clean up utils index
